requirements.txt
# Web framework
Flask>=3.0.0
gunicorn>=21.2.0

# AI extraction
anthropic>=0.40.0
requests>=2.31.0
Pillow>=10.0.0

# Google OAuth (ID token verification)
google-auth>=2.29.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
